+++ plugins/postcss-logical/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Logical Properties
 
+### Unreleased (major)
+
+- Updated: Support for Node v14+ (major).
+
 ### 5.0.4 (February 5, 2022)
 
 - Improved `es module` and `commonjs` compatibility