+++ plugins/postcss-logical/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased (major)
 
 - Updated: Support for Node v14+ (major).
+- Added: TypeScript support.
 - Added: Support for `block-size`, `inline-size` and `caption-side`.
 - Removed: Support for the `logical` keyword within `margin`, `padding`, `border-width`, `border-style`, `border-color` properties. (major)
 - Updated: The plugin now requires block and inline direction to be configured (uses `top-to-bottom` and `left-to-right` respectively) and no longer introduces `:dir` (major).